wip: error with isRef
