+++ packages/c-button/examples/index.ts
@@ -1,0 +1,8 @@
+export * as AsIconButton from './as-icon-button.vue'
+export * as BaseButton from './base-button.vue'
+export * as WithAttachedButtons from './with-attached-buttons.vue'
+export * as WithButtonGroup from '../examples/with-button-group.vue'
+export * as WithButtonSize from '../examples/with-button-size.vue'
+export * as withButtonIcon from '../examples/with-button-icon.vue'
+export * as WithButtonVariants from '../examples/with-button-variants.vue'
+export * as WithLoadingStatus from '../examples/with-loading-status.vue'
