--- packages/c-button/tests/c-button.cy.ts
@@ -1,19 +1,0 @@
-import { h } from 'vue'
-import { mount } from '@cypress/vue'
-import theme from '@chakra-ui/vue-theme'
-import { CButton } from '../src'
-
-it('should render properly', () => {
-  mount(CButton, {
-    global: {
-      provide: {
-        $chakraTheme: theme,
-        $chakraColorMode: 'light',
-        $chakraIcons: {},
-      },
-    },
-    slots: {
-      default: 'Button',
-    },
-  })
-})
