+++ packages/c-button/tests/c-button.test.tsx
@@ -1,0 +1,15 @@
+import * as Examples from '../examples'
+import { h } from 'vue'
+
+
+describe('Button Examples', () => {
+  it('works', () => {
+    expect(true).to.be.true
+  })
+
+  Object.entries(Examples).map(([name, example]) => {
+    it(`renders ${name} successfully`, () => {
+      cy.mount(example.default).checkA11y()
+    })
+  })
+})
