+++ packages/c-popper/examples/index.ts
@@ -1,0 +1,2 @@
+export * as PopperV2Example from './popper-v2-example.vue'
+export * as PopperWithTransition from './popper-with-transition.vue'
